+++ NEWS.md
@@ -1,3 +1,10 @@
+# tsibble 0.9.1
+
+* Fixed bugs when `index_by()` + `summarise()` with grouped factors. (#197)
+* Fixed `select()` when removing redundant key. (#196)
+* Added a new message in `new_data.grouped_ts()` to inform "grouping structure is ignored". (#193)
+* `vec_slice()` updates tsibble data attributes.
+
 # tsibble 0.9.0
 
 This release uses the **vctrs** package internally in {tsibble}.